Make error messages more useful
Just convert the list by joining the elements with space. It may not be
perfect since it may explode in cases we have spaces, but it's better
than giving a list that you can't do anything with it.

By adding the error output as a command one can manually build that
single compilation unit to understand better the error and even switch
to another compile to get a better error message.